backend/requirements.txt
Django>=5.2,<5.3
djangorestframework>=3.16.1,<3.17
psycopg2>=2.9.11,<3.0
drf-spectacular>=0.29.0,<0.30
djangorestframework-simplejwt>=5.5.1,<5.6
setuptools>=75,<82